fix: add support for all valid Argon2 type variants
This commit changes the behavior of the Argon2Hasher identify method to
correctly identify and signal support for all valid Argon2 type variants
which at the time of writing include 'argon2id', 'argon2d', and
'argon2i'.

This commit also enhances the hash validation mechanism to move from
only checking the hash type against a hard-coded string to validating
the complete hash against a comprehensive regex pattern and verifying
the type variant against the `argon2-cffi` backend library to ensure the
user-supplied hash is a valid and well-formed Argon2 hash.

Resolves #21.